pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent", "src/gateway", "src/security", "src/sandbox", "src/cli", "src/exceptions"]

[project]
name = "clotho"
version = "0.1.0"
description = "A CLI-based coding agent"
license = {text = "MIT"}
authors = [{name = "Alec Sirkin"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "docker>=7.1.0",
    "fastapi>=0.115.0",
    "ollama>=0.6.1",
    "openai>=2.21.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "rich>=13.0.0",
    "setproctitle>=1.3.7",
]

[project.scripts]
clotho = "cli.main:main"
